+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-## [v7.2.0] - Unreleased
+## [v7.2.0] - 2020-11-10
 
 ### Changed
 - Do not try activating the module if its already activated [PR-25](https://github.com/OXID-eSales/testing_library/pull/25)